requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
blessed==1.20.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
croniter==2.0.1
crontab==1.0.1
cryptography==41.0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-compat==1.0.15
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-import-export==3.3.5
django-picklefield==3.1
django-redis==5.4.0
django-rest-passwordreset==1.3.0
django-rq==2.10.1
django-tasks-scheduler==1.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
freezegun==1.4.0
future==0.18.3
idna==3.6
jinxed==1.2.1
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
pyee==11.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rq==1.15.1
rq-scheduler==0.13.1
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
times==0.7
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
